pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Mariana Lishchynska",email = "lishchynska.m@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram (>=3.19.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "google-genai (>=1.10.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
